% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDrillClass.R
\docType{class}
\name{drill-class}
\alias{drill-class}
\title{Class \code{drill} defines a Apache Drill driver class.}
\description{
Class \code{drill} defines a Apache Drill driver class.
}
\section{Slots}{

\describe{
\item{\code{host}}{A string denoting the Apache Drill cluster/node host}

\item{\code{port}}{A numeric denoting the Apache Drill cluster/node port}
}}

