% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_choices.R
\name{format_choices}
\alias{format_choices}
\title{Format question/answer strings}
\usage{
format_choices(choice_str)
}
\description{
Question/answer character strings can be identified to be
latex markup or an image or just plain text.
}
