% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDrillGeneric.R
\docType{methods}
\name{rd_active}
\alias{rd_active}
\alias{rd_active,drill-method}
\title{To show whether the connection is active.}
\usage{
rd_active(object)

\S4method{rd_active}{drill}(object)
}
\arguments{
\item{object}{the RDrill object}
}
\description{
To show whether the connection is active.
}
\examples{
#rd_active(R_DRILL_OBJECT) #please replace 'R_DRILL_OBJECT' with your RDrill object.
}
\keyword{RDrillGeneric}

