% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDrillGeneric.R
\docType{methods}
\name{rd_metrics}
\alias{rd_metrics}
\alias{rd_metrics,drill-method}
\title{Show the information of the memory metrics.}
\usage{
rd_metrics(object)

\S4method{rd_metrics}{drill}(object)
}
\arguments{
\item{object}{the RDrill object}
}
\description{
Show the information of the memory metrics.
}
\examples{
#rd_metrics(R_DRILL_OBJECT) #please replace 'R_DRILL_OBJECT' with your RDrill object.
}
\keyword{RDrillGeneric}

