% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDrillGeneric.R
\docType{methods}
\name{rd_options}
\alias{rd_options}
\alias{rd_options,drill-method}
\title{Show the information of the system and session options such as name and data type and so forth.}
\usage{
rd_options(object)

\S4method{rd_options}{drill}(object)
}
\arguments{
\item{object}{the RDrill object}
}
\description{
Show the information of the system and session options such as name and data type and so forth.
}
\examples{
#rd_options(R_DRILL_OBJECT) #please replace 'R_DRILL_OBJECT' with your RDrill object.
}
\keyword{RDrillGeneric}

