% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDrillGeneric.R
\docType{methods}
\name{rd_profiles}
\alias{rd_profiles}
\alias{rd_profiles,drill-method}
\title{Show the information about the running and completed queries.}
\usage{
rd_profiles(object)

\S4method{rd_profiles}{drill}(object)
}
\arguments{
\item{object}{the RDrill object}
}
\description{
Show the information about the running and completed queries.
}
\examples{
#rd_profiles(R_DRILL_OBJECT) #please replace 'R_DRILL_OBJECT' with your RDrill object.
}
\keyword{RDrillGeneric}

