% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDrillGeneric.R
\docType{methods}
\name{rd_query}
\alias{rd_query}
\alias{rd_query,drill-method}
\title{Submit a SQL query and return results as a data frame.}
\usage{
rd_query(object, sql)

\S4method{rd_query}{drill}(object, sql)
}
\arguments{
\item{object}{the RDrill object}

\item{sql}{the query submitted}
}
\description{
Submit a SQL query and return results as a data frame.
}
\examples{
#rd_query(R_DRILL_OBJECT,SQL_QUERY) #please replace 'R_DRILL_OBJECT' with your RDrill object.
}
\keyword{RDrillGeneric}

