% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDrillGeneric.R
\docType{methods}
\name{rd_stats}
\alias{rd_stats}
\alias{rd_stats,drill-method}
\title{Show the information of Drillbits, their ports and hosts.}
\usage{
rd_stats(object)

\S4method{rd_stats}{drill}(object)
}
\arguments{
\item{object}{the RDrill object}
}
\description{
Show the information of Drillbits, their ports and hosts.
}
\examples{
#rd_stats(R_DRILL_OBJECT) #please replace 'R_DRILL_OBJECT' with your RDrill object.
}
\keyword{RDrillGeneric}

