% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDrillGeneric.R
\docType{methods}
\name{rd_storages}
\alias{rd_storages}
\alias{rd_storages,drill-method}
\title{Show the configurations of the storage plugins.}
\usage{
rd_storages(object)

\S4method{rd_storages}{drill}(object)
}
\arguments{
\item{object}{the RDrill object}
}
\description{
Show the configurations of the storage plugins.
}
\examples{
#rd_storages(R_DRILL_OBJECT) #please replace 'R_DRILL_OBJECT' with your RDrill object.
}
\keyword{RDrillGeneric}

