% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDrillGeneric.R
\docType{methods}
\name{rd_threads}
\alias{rd_threads}
\alias{rd_threads,drill-method}
\title{Show the information of the threads status.}
\usage{
rd_threads(object)

\S4method{rd_threads}{drill}(object)
}
\arguments{
\item{object}{the RDrill object}
}
\description{
Show the information of the threads status.
}
\examples{
#rd_threads(R_DRILL_OBJECT) #please replace 'R_DRILL_OBJECT' with your RDrill object.
}
\keyword{RDrillGeneric}

